+++ console/console-init/ui/src/graphql-module/queries/Address.ts
@@ -225,6 +225,7 @@ const RETURN_ADDRESS_DETAIL = (
           }
           spec {
             address
+            type
             topic
             plan {
               spec {