+++ systemtests/src/test/java/io/enmasse/systemtest/shared/standard/web/ChromeConsoleTest.java
@@ -207,8 +207,8 @@ public class ChromeConsoleTest extends ConsoleTest implements ITestSharedStandar
 
     @Test
     @ExternalClients
-    void testEmptyLinkPage() throws Exception {
-        doTestEmptyLinkPage(getSharedAddressSpace(), TestInfo.getInstance().getActualTest());
+    void testGoneAwayPageAfterConnectionClose() throws Exception {
+        doTestGoneAwayPageAfterConnectionClose(getSharedAddressSpace(), TestInfo.getInstance().getActualTest());
     }
 
     @Test()