+++ systemtests/src/test/java/io/enmasse/systemtest/shared/standard/web/FirefoxConsoleTest.java
@@ -213,6 +213,13 @@ public class FirefoxConsoleTest extends ConsoleTest implements ITestSharedStanda
     }
 
     @Test
+    @ExternalClients
+    @Tag(ACCEPTANCE)
+    void testConnectionClose() throws Exception {
+        doTestConnectionClose(getSharedAddressSpace());
+    }
+
+    @Test
     void testEditAddress() throws Exception {
         doTestEditAddress(getSharedAddressSpace(), new AddressBuilder()
                         .withNewMetadata()
@@ -318,8 +325,8 @@ public class FirefoxConsoleTest extends ConsoleTest implements ITestSharedStanda
 
     @Test
     @ExternalClients
-    void testEmptyLinkPage() throws Exception {
-        doTestEmptyLinkPage(getSharedAddressSpace(), TestInfo.getInstance().getActualTest());
+    void testGoneAwayPageAfterConnectionClose() throws Exception {
+        doTestGoneAwayPageAfterConnectionClose(getSharedAddressSpace(), TestInfo.getInstance().getActualTest());
     }
 
     @Test()